requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
anthropic==0.25.0
apscheduler==3.10.4
aiosqlite==0.20.0
python-dotenv==1.0.0
